Move formatting function to only module that uses it
